fix/chat기능 (#38)

+++ src/main/java/com/ureca/miniproject/chat/controller/ChatController.java
@@ -6,6 +6,7 @@ import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.messaging.handler.annotation.SendTo;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Controller;
 
 import com.ureca.miniproject.chat.dto.ChatMessage;
@@ -20,23 +21,45 @@ public class ChatController {
 
     private final ChatRoomUserTool userRepo;
     private final StateManager stateManager;
-
+    private final SimpMessagingTemplate messagingTemplate;
     @MessageMapping("/chat.send/{roomId}")
     @SendTo("/topic/chat/{roomId}")
     public ChatMessage send(@Payload ChatMessage message, @DestinationVariable("roomId") String roomId, Principal principal) {
 
         String username = principal.getName();
         message.setSender(username);
-
+        
         switch (message.getType()) {
-            case ENTER -> {
-                userRepo.addUser(roomId, username);
-                message.setMessage(username + "님이 입장하셨습니다.");
-            }
+	        case ENTER -> {
+	
+	            if (!userRepo.isUserInRoom(roomId, username)) {
+	                userRepo.addUser(roomId, username);
+	                message.setMessage(username + "님이 입장하셨습니다.");
+	            } else {
+	                message.setMessage(null);
+	            }
+	
+	            message.setParticipants(userRepo.getUsers(roomId));
+	        }
+
+
+
             case LEAVE -> {
-                userRepo.removeUser(roomId, username);
-                message.setMessage(username + "님이 퇴장하셨습니다.");
+                new Thread(() -> {
+                    try {
+                        Thread.sleep(3000); 
+                        if (!userRepo.isUserInRoom(roomId, username)) {
+                            userRepo.removeUser(roomId, username);
+                            message.setMessage(username + "님이 퇴장하셨습니다.");
+                            messagingTemplate.convertAndSend("/topic/chat/" + roomId, message);
+                        }
+                    } catch (InterruptedException e) {
+                        Thread.currentThread().interrupt();
+                    }
+                }).start();
+                return null; 
             }
+
             case START_TIME -> {
                 message.setParticipants(userRepo.getUsers(roomId));
                 stateManager.startDebate(roomId, message);
